refactor: card-shell
