+++ src/pages/former/former-page.tsx
@@ -1,5 +1,5 @@
 import { CardCreator } from '@/components/card-creator';
-import CardShell from '@/components/card-shell';
+import { CardShell } from '@/components/card-shell';
 import { AppContext, CardData } from '@/providers';
 import React from 'react';
 import './former-page.css';
