+++ src/pages/main/main-page.tsx
@@ -1,4 +1,4 @@
-import CardShell from '@/components/card-shell';
+import { CardShell } from '@/components/card-shell';
 import SearchBar from '@/components/searchbar';
 import { AppContext, CardData } from '@/providers';
 import React from 'react';
